Updated API and readme
