+++ test/audience.test.ts
@@ -12,23 +12,19 @@ describe('Test edkt audience matching', () => {
   beforeAll(async () => {
     localStorage.clear();
     // add one initial view
-    await edkt.run(
-      {
-        pageFeatureGetters: [sportPageFeatureGetter],
-        audienceDefinitions: [sportInterestAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [sportPageFeatureGetter],
+      audienceDefinitions: [sportInterestAudience],
+      omitGdprConsent: true,
+    });
   });
 
   it('First run -> add page view but do not match', async () => {
-    await edkt.run(
-      {
-        pageFeatureGetters: [sportPageFeatureGetter],
-        audienceDefinitions: [sportInterestAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [sportPageFeatureGetter],
+      audienceDefinitions: [sportInterestAudience],
+      omitGdprConsent: true,
+    });
 
     const edktPageViews = JSON.parse(
       localStorage.getItem('edkt_page_views') || '[]'
@@ -43,13 +39,11 @@ describe('Test edkt audience matching', () => {
   });
 
   it('Second run -> add another page view & match', async () => {
-    await edkt.run(
-      {
-        pageFeatureGetters: [sportPageFeatureGetter],
-        audienceDefinitions: [sportInterestAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [sportPageFeatureGetter],
+      audienceDefinitions: [sportInterestAudience],
+      omitGdprConsent: true,
+    });
 
     const edktPageViews = JSON.parse(
       localStorage.getItem('edkt_page_views') || '[]'
@@ -64,13 +58,11 @@ describe('Test edkt audience matching', () => {
   });
 
   it('Third run -> add another page view & match', async () => {
-    await edkt.run(
-      {
-        pageFeatureGetters: [sportPageFeatureGetter],
-        audienceDefinitions: [sportInterestAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [sportPageFeatureGetter],
+      audienceDefinitions: [sportInterestAudience],
+      omitGdprConsent: true,
+    });
 
     const edktPageViews = JSON.parse(
       localStorage.getItem('edkt_page_views') || '[]'
