+++ test/run.test.ts
@@ -86,13 +86,11 @@ describe('Test basic edkt run', () => {
   it('does not match with one sport page view', async () => {
     setUpLocalStorage(ONE_SPORTS_PAGE_VIEW);
 
-    await edkt.run(
-      {
-        pageFeatureGetters: [sportPageFeatureGetter],
-        audienceDefinitions: [sportAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [sportPageFeatureGetter],
+      audienceDefinitions: [sportAudience],
+      omitGdprConsent: true,
+    });
 
     const edktPageViews = JSON.parse(
       localStorage.getItem('edkt_page_views') || '[]'
@@ -112,13 +110,11 @@ describe('Test basic edkt run', () => {
   it('does match with two sport page view', async () => {
     setUpLocalStorage(TWO_SPORTS_PAGE_VIEW);
 
-    await edkt.run(
-      {
-        pageFeatureGetters: [sportPageFeatureGetter],
-        audienceDefinitions: [sportAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [sportPageFeatureGetter],
+      audienceDefinitions: [sportAudience],
+      omitGdprConsent: true,
+    });
 
     const edktPageViews = JSON.parse(
       localStorage.getItem('edkt_page_views') || '[]'
@@ -140,13 +136,11 @@ describe('Test look back edkt run', () => {
   it('does match with lookBack set to 0 with two demo page view at any point in the past', async () => {
     setUpLocalStorage(LOOK_BACK_INFINITY_PAGE_VIEW);
 
-    await edkt.run(
-      {
-        pageFeatureGetters: [lookBackPageFeatureGetter],
-        audienceDefinitions: [lookBackInfinityAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [lookBackPageFeatureGetter],
+      audienceDefinitions: [lookBackInfinityAudience],
+      omitGdprConsent: true,
+    });
 
     const edktMatchedAudiences = edkt.getMatchedAudiences();
     expect(edktMatchedAudiences.length).toEqual(1);
@@ -156,13 +150,11 @@ describe('Test look back edkt run', () => {
   it('does match with lookBack set to 2 with two blank page view within look back period', async () => {
     setUpLocalStorage(LOOK_BACK_PAGE_VIEW);
 
-    await edkt.run(
-      {
-        pageFeatureGetters: [lookBackPageFeatureGetter],
-        audienceDefinitions: [lookBackAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [lookBackPageFeatureGetter],
+      audienceDefinitions: [lookBackAudience],
+      omitGdprConsent: true,
+    });
 
     const edktMatchedAudiences = edkt.getMatchedAudiences();
     expect(edktMatchedAudiences.length).toEqual(1);
@@ -172,13 +164,11 @@ describe('Test look back edkt run', () => {
   it('does not match with lookBack set to 2 with two blank page view outside look back period', async () => {
     setUpLocalStorage(LOOK_BACK_INFINITY_PAGE_VIEW);
 
-    await edkt.run(
-      {
-        pageFeatureGetters: [lookBackPageFeatureGetter],
-        audienceDefinitions: [lookBackAudience],
-      },
-      true
-    );
+    await edkt.run({
+      pageFeatureGetters: [lookBackPageFeatureGetter],
+      audienceDefinitions: [lookBackAudience],
+      omitGdprConsent: true,
+    });
 
     const edktMatchedAudiences = edkt.getMatchedAudiences();
     expect(edktMatchedAudiences.length).toEqual(0);
